refactor(routes): extract agents-crud.ts from agents.ts (Menet 5)
Moves all CRUD and utility handlers out of agents.ts into a new
tryHandleAgentsCrud(ctx, webDir) function: list, activity, model-suggest,
create, avatar upload/GET, security GET/PUT, team graph, team GET/PUT,
export-all, export, import, voice-config, and master :name GET/PUT/DELETE.

agents.ts is now a pure delegation shell (~20 lines):
  tryHandleAgentsModels -> tryHandleAgentsChannels
  -> tryHandleAgentsProcess -> tryHandleAgentsCrud

Master :name handler placed last to avoid shadowing specific-suffix routes.
webDir threaded through agents.ts -> tryHandleAgentsCrud for avatar gallery.

tsc --noEmit: EXIT 0 | curl GET /api/agents: 200
vitest channel-stability-contract + main-agent-detail-guards: 24/24 passed
